move cli to _cli
